fix: 修复AI模块编译警告和类型安全问题
- 修复ModelMetrics的Default derive问题（Instant无法派生Default）
- 修复类型转换警告：u32 → u64避免溢出
- 清理未使用的导入（OwnedSemaphorePermit, JoinHandle, mpsc）
- 修复借用检查问题：task_results正确传递到worker_loop
- 添加显式类型注解避免推断歧义

+++ src/ai_memory_pool.rs
@@ -139,18 +139,18 @@ impl MemoryPoolStats {
     pub fn record_allocation(&mut self, time: Duration) {
         self.total_allocations += 1;
         let current_avg = self.average_allocation_time;
-        let count = self.total_allocations as u32;
+        let count = self.total_allocations as u64;
         self.average_allocation_time = Duration::from_nanos(
-            (current_avg.as_nanos() as u64 * (count - 1) + time.as_nanos() as u64) / count as u64
+            (current_avg.as_nanos() as u64 * (count - 1) + time.as_nanos() as u64) / count
         );
     }
 
     pub fn record_deallocation(&mut self, time: Duration) {
         self.total_deallocations += 1;
         let current_avg = self.average_deallocation_time;
-        let count = self.total_deallocations as u32;
+        let count = self.total_deallocations as u64;
         self.average_deallocation_time = Duration::from_nanos(
-            (current_avg.as_nanos() as u64 * (count - 1) + time.as_nanos() as u64) / count as u64
+            (current_avg.as_nanos() as u64 * (count - 1) + time.as_nanos() as u64) / count
         );
     }
 }
